requirements.txt
# Core TTS
kokoro>=0.9.2
soundfile
sounddevice

# PDF & Formats
PyMuPDF
pdfplumber
ebooklib
beautifulsoup4
python-docx

# NLP / Text
nltk
num2words

# GUI
PyQt6>=6.6.0

# System
numpy
